Relayout the function ofthe project & fix warnings
