+++ include/waffle/wm.h
@@ -2,10 +2,10 @@
     #define WM_H_
 
     #include <stdbool.h>
+
     #include <X11/Xlib.h>
     #include <X11/keysym.h>
 
-    #include "debug.h"
     #include "utils.h"
 
 typedef struct {
@@ -24,36 +24,7 @@ typedef struct {
     Window focused_window;
 } wm_state_t;
 
-static inline
-void setup_grab(Display *display, Window root_id)
-{
-    static int grab_event_pointer = (
-        PointerMotionMask
-        | ButtonPressMask
-        | ButtonReleaseMask
-    );
-
-    DEBUG("GEP: %d", grab_event_pointer);
-    XSelectInput(
-        display, root_id,
-        EnterWindowMask | LeaveWindowMask| SubstructureRedirectMask
-    );
-    XGrabButton(
-        display, Button1Mask | Button3Mask, Mod4Mask,
-        root_id, False, grab_event_pointer,
-        GrabModeAsync, GrabModeSync, None, None
-    );
-    XGrabKeyboard(
-        display, root_id, True,
-        GrabModeAsync, GrabModeAsync, CurrentTime
-    );
-}
-
-static inline
-void remove_grab(Display *display, Window root_id)
-{
-    XUngrabButton(display, AnyButton, AnyModifier, root_id);
-    XUngrabKeyboard(display, CurrentTime);
-}
+void setup_grab(Display *display, Window root_id);
+void remove_grab(Display *display, Window root_id);
 
 #endif /* !WM_H_ */
