+++ src/events_debug.c
@@ -7,7 +7,7 @@
 #include "waffle/events.h"
 
 DEBUG_USED
-void debug_mouse_motion(wm_state_t *wm_state, bool mode)
+void debug_mouse_motion(wm_state_t *wm_state DEBUG_USED, bool mode)
 {
     static bool save = true;
 
@@ -23,7 +23,7 @@ void debug_mouse_motion(wm_state_t *wm_state, bool mode)
 }
 
 DEBUG_USED
-void debug_win_rect(Display *display, Window win, bool mode)
+void debug_win_rect(Display *display DEBUG_USED, Window win, bool mode)
 {
     static bool save = true;
     XWindowAttributes attr;
